// dashboard-app/src/app/modules/campaigns/models/campaign-stats.ts
export class CampaignStats {
    id: number;
    campaign_name: string;

    users_show: number;
    users_detail: number;
    users_new: number;
    users_new_monthly: number;
    users_new_one_time:number;

    donation_collected_all: number;
    donation_collected_monthly: number;
    donation_collected_one_time: number;
    donation_average: number;
    donation_median:number;
    donation_goal:number;

}